completed systematic review of all modules

+++ modules/agrvate/main.nf
@@ -29,7 +29,7 @@ process AGRVATE {
     (_meta: Map, fna: Path): Record
 
     stage:
-    stageAs 'input/*', fna
+    stageAs "fna/*", fna
 
     output:
     record(
@@ -59,11 +59,12 @@ process AGRVATE {
     meta.process_name = task.ext.process_name
 
     def is_compressed = fna.getName().endsWith(".gz") ? true : false
-    def fna_name = "${prefix}.fna"
+    def fna_name = fna.getName().replace(".gz", "")
     """
     if [ "${is_compressed}" == "true" ]; then
         gzip -c -d ${fna} > ./${fna_name}
     else
+        # agrvate does not support symlinks
         cp ${fna} ./${fna_name}
     fi
 
@@ -75,7 +76,9 @@ process AGRVATE {
     mv supplemental/${prefix}-summary.tab ./${prefix}.tsv
 
     # Cleanup
-    rm -rf ${fna_name}
+    if [ "${is_compressed}" == "true" ]; then
+        rm -rf ${fna_name}
+    fi
 
     cat <<-END_VERSIONS > versions.yml
     "${task.process}":
