+++ modules/checkm2/predict/main.nf
@@ -36,7 +36,7 @@ process CHECKM2_PREDICT {
 
     input:
     (_meta: Map, fna: Path): Record
-    db                : Path
+    db: Path
 
     output:
     record(
@@ -93,7 +93,9 @@ process CHECKM2_PREDICT {
 
     # Cleanup
     gzip supplemental/protein_files/*.faa
-    rm -rf ${fna_name}
+    if [ "${is_compressed}" == "true" ]; then
+        rm -rf ${fna_name}
+    fi
 
     cat <<-END_VERSIONS > versions.yml
     "${task.process}":
