+++ modules/meningotype/main.nf
@@ -68,7 +68,9 @@ process MENINGOTYPE {
         > ${prefix}.tsv
 
     # Cleanup
-    rm -rf ${fna_name}
+    if [ "${is_compressed}" == "true" ]; then
+        rm -rf ${fna_name}
+    fi
 
     cat <<-END_VERSIONS > versions.yml
     "${task.process}":
