+++ modules/spatyper/main.nf
@@ -34,8 +34,8 @@ process SPATYPER {
 
     input:
     (_meta: Map, fna: Path): Record
-    repeats        : Path?
-    repeat_order   : Path?
+    repeats     : Path?
+    repeat_order: Path?
 
     output:
     record(
@@ -78,7 +78,9 @@ process SPATYPER {
         --output ${prefix}.tsv
 
     # Cleanup
-    rm -rf ${fna_name}
+    if [ "${is_compressed}" == "true" ]; then
+        rm -rf ${fna_name}
+    fi
 
     cat <<-END_VERSIONS > versions.yml
     "${task.process}":
